added keep-connection functionality fixed encoder for encryped data field + minor fixes

+++ src/main/scala/tserver/utils/FactorizationUtil.scala
@@ -5,9 +5,9 @@ import scala.annotation.tailrec
 object FactorizationUtil {
 
   //copy-paste form stackoverflow
-  def factorize(x: Int): List[Int] = {
+  def factorize(x: Long): List[Long] = {
     @tailrec
-    def foo(x: Int, a: Int = 2, list: List[Int] = Nil): List[Int] = a*a > x match {
+    def foo(x: Long, a: Long = 2, list: List[Long] = Nil): List[Long] = a*a > x match {
       case false if x % a == 0 => foo(x / a, a    , a :: list)
       case false               => foo(x    , a + 1, list)
       case true                => x :: list
